Templated input and output arrays, index type for observations. (#1)
* Refactor ncp handling to account for unsigned indices.
* Throw if the index integer type is too small.

+++ tests/src/TestCore.h
@@ -38,7 +38,7 @@ protected:
         return output;
     }
         
-    size_t nr, nc;
+    int nr, nc;
     std::vector<double> data;
 };
